requirements.txt
# Python Dependencies
# Python 3.11+

# Web framework
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
jinja2>=3.1.0,<4.0
python-multipart
itsdangerous

# HTTP client (Telegram)
httpx>=0.27.0,<1.0

# Configuration
pyyaml>=6.0,<7.0
